backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
httpx>=0.24.1
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
chardet>=4.0.0
lxml>=4.9.0
selenium>=4.0.0
webdriver-manager>=3.8.0
openai==1.12.0
google-generativeai==0.3.2